add different type fields to form
